#32 Add constructor to WorkoutSession for GpsFile. Write test for it.

+++ tdt4140-gr1817/improvify.gpx/src/main/java/tdt4140.gr1817.improvify.gpx/GpsFile.java
@@ -12,9 +12,9 @@ public class GpsFile {
     private static int hoursInDay = 24;
 
     private LocalDateTime startTime;
-    private int averageHeartRate;
-    private int maxHeartRate;
-    private int distanceRun;
+    private float averageHeartRate;
+    private float maxHeartRate;
+    private float distanceRun;
     private int duration; //in seconds
 
     public GpsFile(ArrayList<GpsPoint> pointList) {
